hi vahd
Cock and ball torture From Wikipedia, the free encyclopedia at en.wikipedia.org
Cock and ball torture (CBT) is a sexual activity involving application of pain or constriction to the male genitals. This may involve directly painful activities, such as genital piercing, wax play, genital spanking, squeezing, ball-busting, genital flogging, urethral play, tickle torture, erotic electrostimulation or even kicking. The recipient of such activities may receive direct physical pleasure via masochism, or knowledge that the play is pleasing to a sadistic dominant.
Image: Electrostimulation applied on a penis
Contents: Section 1: In Pornography Section 2: Ball stretcher Section 3: Parachute Section 4: Humbler Section 5: Testicle cuff
Section 1: In pornography
In addition to it’s occasional role in BDSM pornography, Tamakeri (literally Ball kicking) is a separate genre in Japan. One notable actress in tamakeri is Erika Nagai who typically uses her martial arts skills to knee or kick men in the testicles.
Section 2: Ball stretcher A ball stretcher is a sex toy that is fastened around a man in order to elongate the scrotum and provide a feeling of weight pulling the testicles away from the body. While leather stretchers are most common, other models are made of steel rings that are fastened with screws causing additional mildly uncomfortable weight to the wearer. The length of the stretcher may vary from 1-4 inches, and the steel models can weigh as much as five pounds.
Section 3: Parachute
A Parachute is a small collar, usually made from leather, which fastens around the scrotum, and from which weights can be hung. Conical in shape, three or four short chains hanging beneath, to which weights can be attached. Used as part of cock and ball torture within a BDSM relationship, the parachute provides a constant drag, and squeezing effects on the man’s testicles. Moderate weights of 3-5 kg can be suspended, especially during bondage. Smaller weights can be used when the man is free to move, when the swinging effect of the weight can restrict sudden movements, as well as providing a visual stimulus for the dominant partner.
Section 4: Humbler
A humbler is a BDSM physical restraint device used to restrict the movement of a submissive male participant in a BDSM scene. The humbler consists of a testicle cuff device which clamps around the base of the scrotum, mounted in the center of a bar that passes behind the thighs at the base of the buttocks. This forces the wearer to keep his legs forward, as any attempt to to straighten the legs slightly pulls directly on the scrotum, causing from considerable discomfort to extreme pain.
Section 5: Testicle cuff
A testicle cuff is a ring-shaped device around the scrotum between the body and the testicles which when closed does not allow the testicles to slide through it. A common type has two connected cuffs, one around the scrotum and the other around the base of the penis. They are just one of many devices to restrain the male genitalia. A standard padlock may also be locked around the scrotum; without the key it cannot be removed.
Some passive men enjoy the feeling of being "owned", while dominant individuals enjoy the sense of "owning" their partners. Requiring such a man wear testicle cuffs symbolizes that his sexual organs belong to his partner, who may be either male or female. There is a level of humiliation involved, by which they find sexual arousal. The cuffs may even form part of a sexual fetish of the wearer or his partner.
However, these are extreme uses of testicle cuffs. More conventionally, the device pulls down the testicles and keeps them there during stimulation, which has a number of benefits:
Making the penis appear longer. Pulling the testicles down and away from the base of the penis stretches the skin over the base of the penis and pubic bone, exposing the additional inch or so of penile shaft that is normally hidden from view. Improving sexual arousal. While some men may be aroused by the feeling of being "owned", the physical feeling of stretching the ligaments that suspend the testicles has an effect similar to the more common practice of stretching one's legs and pointing the toes. Preventing the testicles from lifting up so far that they become lodged under the skin immediately adjacent to the base of the penis, a condition which can be very uncomfortable, especially if the testicle is then squashed by the slap of skin during thrusting in sexual intercourse. Delaying or intensifying ejaculation by preventing the testicles from rising normally to the "point of no return". It is much harder to reach an orgasm.